# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients/{patientId}/symptom-reports": {
      "post": {
        "operationId": "PatientSymptomReportsController_create",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientSymptomReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientSymptomReportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "404": {
            "description": "Patient, follow-up, or enrollment not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a patient symptom report",
        "tags": [
          "Patient symptom reports"
        ]
      },
      "get": {
        "operationId": "PatientSymptomReportsController_findAll",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientSymptomReportResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List patient symptom reports",
        "tags": [
          "Patient symptom reports"
        ]
      }
    },
    "/psychooncology-appointments": {
      "post": {
        "operationId": "PsychooncologyAppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePsychooncologyAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychooncologyAppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Required agent assignment is missing"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Agents cannot assign appointments to others"
          },
          "404": {
            "description": "Patient, availability slot, volunteer, or agent not found"
          },
          "409": {
            "description": "Availability slot is reserved or volunteer is inactive"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule a psycho-oncology appointment",
        "tags": [
          "psychooncology-appointments"
        ]
      },
      "get": {
        "operationId": "PsychooncologyAppointmentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PsychooncologyAppointmentResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The authenticated volunteer has no profile"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List psycho-oncology appointments",
        "tags": [
          "psychooncology-appointments"
        ]
      }
    },
    "/psychooncology-appointments/{id}": {
      "get": {
        "operationId": "PsychooncologyAppointmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychooncologyAppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The authenticated volunteer has no profile"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Volunteers can only access their own appointments"
          },
          "404": {
            "description": "Appointment not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a psycho-oncology appointment",
        "tags": [
          "psychooncology-appointments"
        ]
      },
      "patch": {
        "operationId": "PsychooncologyAppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePsychooncologyAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychooncologyAppointmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid appointment status transition"
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "403": {
            "description": "Volunteers have limited update permissions"
          },
          "404": {
            "description": "Appointment or availability slot not found"
          },
          "409": {
            "description": "Closed appointments cannot be updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a psycho-oncology appointment",
        "tags": [
          "psychooncology-appointments"
        ]
      }
    },
    "/psychooncology-appointments/{id}/cancel": {
      "patch": {
        "operationId": "PsychooncologyAppointmentsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PsychooncologyAppointmentResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "JWT missing, invalid, or expired"
          },
          "404": {
            "description": "Appointment or availability slot not found"
          },
          "409": {
            "description": "Closed appointments cannot be cancelled"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a psycho-oncology appointment",
        "tags": [
          "psychooncology-appointments"
        ]
      }
    },
...
  },
  "info": {
    "title": "FPC Backend API",
    "description": "FPC backend API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePatientSymptomReportDto": {
        "type": "object",
        "properties": {
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "enrollmentId": {
            "type": "string",
            "format": "uuid"
          },
          "discomfortSeverity": {
            "type": "string",
            "maxLength": 20
          },
          "discomfortDescription": {
            "type": "string"
          },
          "symptomDuration": {
            "type": "string",
            "maxLength": 50
          },
          "symptomFrequency": {
            "type": "string",
            "maxLength": 50
          },
          "isPainPresent": {
            "type": "boolean"
          },
          "painIntensity": {
            "type": "number",
            "minimum": 0,
            "maximum": 10
          },
          "painLocation": {
            "type": "string",
            "maxLength": 255
          },
          "painDescription": {
            "type": "string"
          },
          "hasSoughtMedicalConsultation": {
            "type": "boolean"
          },
          "healthCenterId": {
            "type": "string",
            "format": "uuid"
          },
          "specialty": {
            "type": "string",
            "maxLength": 255
          }
        },
        "required": [
          "followUpId"
        ]
      },
      "PatientSymptomReportResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "enrollmentId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "discomfortSeverity": {
            "type": "string",
            "nullable": true
          },
          "discomfortDescription": {
            "type": "string",
            "nullable": true
          },
          "symptomDuration": {
            "type": "string",
            "nullable": true
          },
          "symptomFrequency": {
            "type": "string",
            "nullable": true
          },
          "isPainPresent": {
            "type": "boolean",
            "nullable": true
          },
          "painIntensity": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 10
          },
          "painLocation": {
            "type": "string",
            "nullable": true
          },
          "painDescription": {
            "type": "string",
            "nullable": true
          },
          "hasSoughtMedicalConsultation": {
            "type": "boolean"
          },
          "healthCenterId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "specialty": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "followUpId",
          "enrollmentId",
          "discomfortSeverity",
          "discomfortDescription",
          "symptomDuration",
          "symptomFrequency",
          "isPainPresent",
          "painIntensity",
          "painLocation",
          "painDescription",
          "hasSoughtMedicalConsultation",
          "healthCenterId",
          "specialty",
          "createdAt"
        ]
      },
      "CreatePsychooncologyAppointmentDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "availabilityId": {
            "type": "string",
            "format": "uuid"
          },
          "agentId": {
            "type": "string",
            "format": "uuid"
          },
          "patientEmail": {
            "type": "string"
          },
          "isAdditionalSession": {
            "type": "boolean"
          },
          "modality": {
            "type": "string",
            "enum": [
              "CALL",
              "VIDEO_CALL"
            ]
          }
        },
        "required": [
          "patientId",
          "availabilityId",
          "modality"
        ]
      },
      "PsychooncologyAppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "volunteerId": {
            "type": "string",
            "format": "uuid"
          },
          "followUpId": {
            "type": "string",
            "format": "uuid"
          },
          "availabilityId": {
            "type": "string",
            "format": "uuid"
          },
          "patientEmail": {
            "type": "string",
            "nullable": true,
            "format": "email"
          },
          "sessionNumber": {
            "type": "number"
          },
          "isAdditionalSession": {
            "type": "boolean"
          },
          "modality": {
            "enum": [
              "CALL",
              "VIDEO_CALL"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_ANSWER"
            ],
            "type": "string"
          },
          "scheduledAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "topicAddressed": {
            "type": "string",
            "nullable": true
          },
          "sessionDetails": {
            "type": "string",
            "nullable": true
          },
          "additionalObservations": {
            "type": "string",
            "nullable": true
          },
          "recommendations": {
            "type": "string",
            "nullable": true
          },
          "referral": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientId",
          "volunteerId",
          "followUpId",
          "availabilityId",
          "patientEmail",
          "sessionNumber",
          "isAdditionalSession",
          "modality",
          "status",
          "scheduledAt",
          "completedAt",
          "topicAddressed",
          "sessionDetails",
          "additionalObservations",
          "recommendations",
          "referral",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePsychooncologyAppointmentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED",
              "NO_ANSWER"
            ]
          },
          "patientEmail": {
            "type": "string"
          },
          "isAdditionalSession": {
            "type": "boolean"
          },
          "topicAddressed": {
            "type": "string"
          },
          "sessionDetails": {
            "type": "string"
          },
          "additionalObservations": {
            "type": "string"
          },
          "recommendations": {
            "type": "string"
          },
          "referral": {
            "type": "string"
          }
        }
      },
...
    }
  }
}